Create index on runs.tasks.active
